adds hook for getting a stat leader
