+++ src/components/fixtures/FixtureAthleteStats.tsx
@@ -1,13 +1,14 @@
 import { IFixture } from "../../types/games"
 import { fixtureSumary } from "../../utils/fixtureUtils"
 import { GameSportAction } from "../../types/boxScore"
-import { useEffect, useMemo, useState } from "react"
+import { useEffect, useMemo } from "react"
 import { Table2 } from "lucide-react"
-import { BoxscoreListRecordItem, BoxscoreTable } from "./boxscore/BoxscoreCategoryList"
+import { BoxscoreTable } from "./boxscore/BoxscoreCategoryList"
 import FixtureTeamSelector from "./boxscore/FixtureTeamSelector"
 import { useBoxscoreFilter } from "../../hooks/fixtures/useBoxscoreFilter"
 import { fixtureAnalytics } from "../../services/analytics/fixtureAnalytics"
 import { useInView } from "react-intersection-observer"
+import { attackBoxscoreList, defenseBoxscoreList, kickingBoxscoreList, disciplineBoxscoreList } from "../../utils/boxScoreUtils"
 
 type Props = {
     fixture: IFixture,
@@ -17,7 +18,7 @@ type Props = {
 export default function FixtureAthleteStats({ fixture, sportActions }: Props) {
 
     const { gameKickedOff } = fixtureSumary(fixture);
-    const [search, setSearch] = useState<string>("");
+    // const [search, setSearch] = useState<string>("");
 
     const { selectedTeamId } = useBoxscoreFilter(fixture);
     const {ref, inView} = useInView({triggerOnce: true});
@@ -94,144 +95,3 @@ export default function FixtureAthleteStats({ fixture, sportActions }: Props) {
     )
 }
 
-function attackBoxscoreList(bs: GameSportAction[], teamId: string): BoxscoreListRecordItem[] {
-    const athleteIds: string[] = [];
-
-    bs.forEach((b) => {
-        if (!athleteIds.includes(b.athlete_id) && b.team_id === teamId) {
-            athleteIds.push(b.athlete_id);
-        }
-    });
-
-    const athleteStats: BoxscoreListRecordItem[] = athleteIds.map((a) => {
-        const stats = bs.filter((b) => b.athlete_id === a);
-
-        const tries = stats.find((b) => b.action === "tries")?.action_count;
-        const points = stats.find((b) => b.action === "points")?.action_count;
-        const passes = stats.find((b) => b.action === "carry_dominant")?.action_count;
-
-        return {
-            stats: [Math.floor(tries ?? 0), Math.floor(points ?? 0), Math.floor(passes ?? 0)],
-            athleteId: a
-        }
-    }).sort((a, b) => {
-        const [, points] = a.stats;
-        const [, bPoints] = b.stats;
-
-        return (bPoints ?? 0) - (points ?? 0)
-    });
-
-
-    return athleteStats;
-
-}
-
-
-function defenseBoxscoreList(bs: GameSportAction[], teamId: string): BoxscoreListRecordItem[] {
-    const athleteIds: string[] = [];
-
-    bs.forEach((b) => {
-        if (!athleteIds.includes(b.athlete_id) && b.team_id === teamId) {
-            athleteIds.push(b.athlete_id);
-        }
-    });
-
-    const athleteStats: BoxscoreListRecordItem[] = athleteIds.map((a) => {
-        const stats = bs.filter((b) => b.athlete_id === a);
-
-        const tackles = stats.find((b) => b.action === "tackles")?.action_count;
-        const dominantTackles = stats.find((b) => b.action === "dominant_tackles")?.action_count;
-        const turnoversWon = stats.find((b) => b.action === "turnover_won")?.action_count;
-
-        return {
-            stats: [Math.floor(tackles ?? 0), Math.floor(dominantTackles ?? 0), Math.floor(turnoversWon ?? 0)],
-            athleteId: a
-        }
-    }).sort((a, b) => {
-        const [tackles] = a.stats;
-        const [bTackles] = b.stats;
-
-        return (bTackles ?? 0) - (tackles ?? 0)
-    }).filter((a) => {
-        const [x, b, c] = a.stats;
-
-        return x > 0;
-    });
-
-
-    return athleteStats;
-
-}
-
-
-function kickingBoxscoreList(bs: GameSportAction[], teamId: string): BoxscoreListRecordItem[] {
-    const athleteIds: string[] = [];
-
-    bs.forEach((b) => {
-        if (!athleteIds.includes(b.athlete_id) && teamId === b.team_id) {
-            athleteIds.push(b.athlete_id);
-        }
-    });
-
-    const athleteStats: BoxscoreListRecordItem[] = athleteIds.map((a) => {
-        const stats = bs.filter((b) => b.athlete_id === a);
-
-        const conversion_goals = stats.find((b) => b.action === "conversion_goals")?.action_count;
-        const drop_goals_scored = stats.find((b) => b.action === "drop_goals_converted")?.action_count;
-        const penalty_goals = stats.find((b) => b.action === "kick_penalty_good")?.action_count;
-
-        return {
-            stats: [Math.floor(conversion_goals ?? 0), Math.floor(drop_goals_scored ?? 0), Math.floor(penalty_goals ?? 0)],
-            athleteId: a
-        }
-    }).sort((a, b) => {
-        const [conversion_goals] = a.stats;
-        const [bConversion_goals] = b.stats;
-
-        return (bConversion_goals ?? 0) - (conversion_goals ?? 0)
-    }).filter((a) => {
-        const [x, b, c] = a.stats;
-
-        return (x > 0) || (b > 0) || (c > 0)
-    });
-
-
-    return athleteStats;
-
-}
-
-function disciplineBoxscoreList(bs: GameSportAction[], teamId: string): BoxscoreListRecordItem[] {
-    const athleteIds: string[] = [];
-
-    bs.forEach((b) => {
-        if (!athleteIds.includes(b.athlete_id) && teamId === b.team_id) {
-            athleteIds.push(b.athlete_id);
-        }
-    });
-
-    const athleteStats: BoxscoreListRecordItem[] = athleteIds.map((a) => {
-        const stats = bs.filter((b) => b.athlete_id === a);
-
-        const red_cards = stats.find((b) => b.action === "red_cards")?.action_count;
-        const yellow_cards = stats.find((b) => b.action === "yellow_cards")?.action_count;
-
-        return {
-            stats: [Math.floor(red_cards ?? 0), Math.floor(yellow_cards ?? 0)],
-            athleteId: a
-        }
-    }).sort((a, b) => {
-        const [redCards] = a.stats;
-        const [bRedCards] = b.stats;
-
-        return (bRedCards ?? 0) - (redCards ?? 0)
-
-    }).filter((a) => {
-        const [x, b] = a.stats;
-
-        return (x > 0) || (b > 0);
-    });
-
-
-    return athleteStats;
-
-}
