+++ src/state/dashboard/gameStory.atoms.ts
@@ -2,7 +2,7 @@ import { atom } from "jotai";
 import { IFixture } from "../../types/games";
 
 /** Holds a boolean value whether a story has been paused or not */
-const isPausedAtom = atom<boolean>(false);
+const isPausedAtom = atom<boolean>(true);
 
 /** Holds the progress of the story slides */
 const progressAtom = atom<number>(0);
